requirements.txt
numpy 
BioPython
keras
